# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Redmouse API documentation.",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api/*",
    "paths": {
...
        "/api/providers": {
            "get": {
                "description": "Get all providers, acept pagination ?page=1\u0026page_provider=10",
                "tags": [
                    "Providers"
                ],
                "summary": "Get all providers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Provider"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "description": "Create provider, ej. {\"description\": \"new provider name\"}",
                "tags": [
                    "Providers"
                ],
                "summary": "Create provider",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Provider"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/providers/{id}": {
            "get": {
                "description": "Find provider",
                "tags": [
                    "Providers"
                ],
                "summary": "Find provider",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Provider"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "description": "Update provider ej. {\"description\": \"update provider name\"}",
                "tags": [
                    "Providers"
                ],
                "summary": "Update provider",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Provider"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete provider",
                "tags": [
                    "Providers"
                ],
                "summary": "Delete provider",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Provider"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/sizes": {
            "get": {
                "description": "Get all sizes, acept pagination ?page=1\u0026page_size=10",
                "tags": [
                    "Sizes"
                ],
                "summary": "Get all sizes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Size"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "description": "Create size, ej. {\"description\": \"new size name\"}",
                "tags": [
                    "Sizes"
                ],
                "summary": "Create size",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Size"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/sizes/{id}": {
            "get": {
                "description": "Find size",
                "tags": [
                    "Sizes"
                ],
                "summary": "Find size",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Size"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "description": "Update size ej. {\"description\": \"update size name\"}",
                "tags": [
                    "Sizes"
                ],
                "summary": "Update size",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Size"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete size",
                "tags": [
                    "Sizes"
                ],
                "summary": "Delete size",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Size"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Provider": {
            "type": "object",
            "required": [
                "email1",
                "names"
            ],
            "properties": {
                "adress": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email1": {
                    "type": "string"
                },
                "email2": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "names": {
                    "type": "string"
                },
                "phone1": {
                    "type": "string"
                },
                "phone2": {
                    "type": "string"
                },
                "rfc": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Size": {
            "type": "object",
            "required": [
                "description"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}